+++ src/components/forms/SubscriptionForm.jsx
@@ -7,16 +7,16 @@ import { faSpinner } from '@fortawesome/free-solid-svg-icons';
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
 import { navigate } from 'gatsby';
 
-import { isValidEmail, FORM_STATES } from '../helpers';
+import { isValidEmail, FORM_STATES } from '../../helpers';
 
-import { Button } from './Button';
+import { Button } from '../Button';
 import { signupOnMixpanel, trackOnMixpanel } from './lib';
 
 const { ERROR, IDLE, INVALID, LOADING } = FORM_STATES;
 
 declare type FormStatus = {| status: 'idle' | 'loading' | 'invalid' | 'error' |};
 
-export class SignupForm extends Component<
+export class SubscriptionForm extends Component<
   {| i18n: I18n, toggle?: () => void, trackingEvent: string |},
   { email: string, ...FormStatus }
 > {
