+++ src/components/forms/index.js
@@ -1,0 +1,4 @@
+// @flow
+
+export * from './RequestDemoForm';
+export * from './SubscriptionForm';
