+++ src/components/lib/index.js
@@ -1,7 +1,5 @@
 // @flow
 
-export * from './signup';
-export * from './demoAccess';
 export * from './textHelpers';
 export * from './useModal';
 export * from './toggleOverlay';
